make Header stateless
